+++ ciScripts/buildPlugin.sh
@@ -1,0 +1,7 @@
+#!/bin/bash
+
+# Pre Build
+./gradlew patchPluginXml
+
+# Build
+./gradlew buildPlugin